refactor: unify supabase client imports
