+++ app/api/generate/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server'
 import OpenAI from 'openai'
-import { supabase } from '@/lib/supabase'
+import { supabase } from '@/lib/supabase-client'
 
 const openai = new OpenAI({
   apiKey: process.env.OPENAI_API_KEY,
