+++ hooks/useUser.tsx
@@ -1,6 +1,6 @@
 ﻿import { useEffect, useState } from 'react'
 import { User } from '@supabase/supabase-js'
-import { supabase } from '@/lib/supabase-Client'
+import { supabase } from '@/lib/supabase-client'
 
 export function useUser() {
     const [user, setUser] = useState<User | null>(null)
